% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iemand-package.R
\docType{package}
\name{iemand-package}
\alias{iemand}
\alias{iemand-package}
\title{iemand: Create Fake People}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/HanjoStudy/satrday_pkg_dev}
  \item Report bugs at \url{https://github.com/HanjoStudy/satrday_pkg_dev/issues}
}

}
\author{
\strong{Maintainer}: Hanjo Odendaal \email{first.last@example.com}

}
\keyword{internal}
